Refactor Details Widget logic
The details widget now takes the information from the location widget. This is because the location widget makes the quality request and has all the scores regarding each location. Without this the details widget would only have the information embedded in the marker, and would have to make another request to the quality service to get all the stats (which are not yet needed, but will be soon)